no message
svn path=/trunk/boinc/; revision=654

+++ client/net_xfer.h
@@ -48,7 +48,7 @@ public:
     char hostname[256];
     int port;
     int blocksize;
-    int nbytes_xfered;        // total number of bytes transferred
+    double nbytes_xfered;        // total number of bytes transferred
 
     void init(char* host, int port, int blocksize);
     int open_server();
